Extract deepFreeze function to mock file
